Lift restriction on comics in ComicsKingdom
